restructuring. splitted postprocessing and training
