--- conf/base/parameters/data_science.yml
@@ -1,7 +1,0 @@
-data_science:
-  classes: ${classes}
-  number_of_features: 8
-  seed: 1337
-  fw_select: PyTorch # PyTorch | TensorFlow
-  training:
-    epochs: 800
