+++ src/qc_data_preparation/pipelines/training/__init__.py
@@ -1,0 +1,10 @@
+"""
+This is a boilerplate pipeline 'mnist_processing'
+generated using Kedro 0.18.1
+"""
+
+from .pipeline import create_pipeline
+
+__all__ = ["create_pipeline"]
+
+__version__ = "0.1"
